Fix the 'famous' getName() on null bug
I don't even know why this is happening. But it seems to happen if a player's spawnposition is in a level wich was previously loaded, but then unloaded. (Then the next autosave gets called => save() gets a level wich is closed therefore the provider has been (probably even destructed) and becomes null wich triggers this getName() on null. 
We should consider checking on all level get functions that are using the provider variable to check if the provider is not null (level not closed) for less plugin crashes)