Fix #1598
There was a race between the DisconnectPacket (0x05) and the ClientDisconnectPacket (0x15) because both were sent about at the same moment. This commit will change how the 0x15 is sent: as a RELIABLE_ORDERED . In this way the client will receive the 0x05 before since it is already sent as an UNRELIABLE (there are no queues both client&server side and it is sent immediatly) . The 0x15 will be sent but ignored by the client because the client will close the connection after the 0x05 but to avoid breaking things i left it. read the pullrequest for a better explanation.